Rename all "settings" to "preferences" and delete useless bootcamp layouts.
