+++ app/src/main/java/com/epfl/drawyourpath/preferences/PreferencesFragment.kt
@@ -1,10 +1,10 @@
-package com.epfl.drawyourpath.settings
+package com.epfl.drawyourpath.preferences
 
 import android.os.Bundle
 import androidx.preference.PreferenceFragmentCompat
 import com.epfl.drawyourpath.R
 
-class SettingsFragment : PreferenceFragmentCompat() {
+class PreferencesFragment : PreferenceFragmentCompat() {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         setPreferencesFromResource(R.xml.preferences, rootKey)
     }
