--- app/src/main/res/layout/activity_bootcamp_google_auth.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_gravity="center">
-
-    <LinearLayout
-        android:layout_width="333dp"
-        android:layout_height="429dp"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <Button
-            android:id="@+id/bc_google_signin"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Sign In With Google" />
-
-        <Button
-            android:id="@+id/bc_google_signout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Sign Out" />
-
-        <TextView
-            android:id="@+id/sign_in_status"
-            android:layout_width="match_parent"
-            android:layout_height="26dp"
-            android:gravity="center"
-            android:text="Not Signed" />
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
