+++ app/src/main/res/layout/fragment_preferences.xml
@@ -9,7 +9,7 @@
         android:id="@+id/settingsTextView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/settings_menu_item"
+        android:text="@string/preferences_menu_item"
         android:textSize="34sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
