+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -25,8 +25,8 @@
         android:title="@string/history_menu_item" />
     <!-- Settings -->
     <item
-        android:id="@+id/settings_menu_item"
+        android:id="@+id/preferences_menu_item"
         android:enabled="true"
         android:icon="@drawable/ic_settings"
-        android:title="@string/settings_menu_item" />
+        android:title="@string/preferences_menu_item" />
 </menu>
