+++ app/src/main/res/values/strings.xml
@@ -8,7 +8,6 @@
     <string name="set_button_firebase">SET</string>
     <string name="profile_menu_item">Profile</string>
     <string name="draw_menu_item">Draw</string>
-    <string name="settings_menu_item">Settings</string>
     <string name="history_menu_item">History</string>
     <string name="community_menu_item">Community</string>
     <string name="friends_menu_item">Friends</string>
@@ -23,4 +22,5 @@
     <string name="confirm_new_password">Confirm new password:</string>
     <string name="cancel">Cancel</string>
     <string name="apply">Apply</string>
+    <string name="preferences_menu_item">Preferences</string>
 </resources>
